use size_t with strlen instead of casting randomly to ints
git-svn-id: https://urjtag.svn.sourceforge.net/svnroot/urjtag/trunk@1698 b68d4a1b-bc3d-0410-92ed-d4ac073336b7

+++ urjtag/src/cmd/cmd_pod.c
@@ -68,7 +68,7 @@ cmd_pod_run (urj_chain_t *chain, char *params[])
         }
 
         urj_pod_sigsel_t it = URJ_POD_CS_NONE;
-        int n = strlen (params[j]);
+        size_t n = strlen (params[j]);
 
         if (n > 4 && (strncasecmp (params[j], "tck", 3) == 0))
             it = URJ_POD_CS_TCK;
